+++ Changelog.md
@@ -16,7 +16,8 @@
 * Parsing `DocxParagraph#sdt`
 * Parsing `DocumentStructure#sdt`
 * Parsing `Presentation#relationships`
-* Parsing `Extension#sparkline_groups` and `SparklineGroup#type`
+* Parsing `Extension#sparkline_groups`
+* Parsing `SparklineGroup#type`, `SparklineGroup#line_weight`
 
 ### Refactor
 * Change interface of `XlsxDrawing#graphic_frame`